+++ src/main/java/org/truffleruby/options/Options.java
@@ -103,6 +103,8 @@ public class Options {
     public final boolean CEXTS;
     /** --cexts-lock=true */
     public final boolean CEXT_LOCK;
+    /** --cexts-prepend-toolchain-to-path=true */
+    public final boolean CEXTS_PREPEND_TOOLCHAIN_TO_PATH;
     /** --keep-handles-alive=false */
     public final boolean CEXTS_KEEP_HANDLES_ALIVE;
     /** --options-log=false */
@@ -239,6 +241,7 @@ public class Options {
         BACKTRACE_ON_RESCUE = options.get(OptionsCatalog.BACKTRACE_ON_RESCUE_KEY);
         CEXTS = options.get(OptionsCatalog.CEXTS_KEY);
         CEXT_LOCK = options.get(OptionsCatalog.CEXT_LOCK_KEY);
+        CEXTS_PREPEND_TOOLCHAIN_TO_PATH = options.get(OptionsCatalog.CEXTS_PREPEND_TOOLCHAIN_TO_PATH_KEY);
         CEXTS_KEEP_HANDLES_ALIVE = options.get(OptionsCatalog.CEXTS_KEEP_HANDLES_ALIVE_KEY);
         OPTIONS_LOG = options.get(OptionsCatalog.OPTIONS_LOG_KEY);
         LOG_LOAD = options.get(OptionsCatalog.LOG_LOAD_KEY);
@@ -370,6 +373,8 @@ public class Options {
                 return CEXTS;
             case "ruby.cexts-lock":
                 return CEXT_LOCK;
+            case "ruby.cexts-prepend-toolchain-to-path":
+                return CEXTS_PREPEND_TOOLCHAIN_TO_PATH;
             case "ruby.keep-handles-alive":
                 return CEXTS_KEEP_HANDLES_ALIVE;
             case "ruby.options-log":